+++ tests/Ports/UnixTest.php
@@ -6,6 +6,7 @@ namespace Tests\Innmind\OperatingSystem\Ports;
 use Innmind\OperatingSystem\{
     Ports\Unix,
     Ports,
+    Config,
 };
 use Innmind\Socket\{
     Internet\Transport,
@@ -19,19 +20,19 @@ class UnixTest extends TestCase
 {
     public function testInterface()
     {
-        $this->assertInstanceOf(Ports::class, Unix::of());
+        $this->assertInstanceOf(Ports::class, Unix::of(Config::of()));
     }
 
     public function testOpen()
     {
-        $ports = Unix::of();
+        $ports = Unix::of(Config::of());
 
         $socket = $ports->open(
             Transport::tlsv12(),
             IPv4::localhost(),
             Port::of(1234),
         )->match(
-            static fn($server) => $server,
+            static fn($server) => $server->unwrap(),
             static fn() => null,
         );